requirements.txt
agno>=2.3.0
openai>=1.0.0
streamlit>=1.30.0
requests>=2.31.0
httpx>=0.27.0
python-dotenv>=1.0.0
reportlab>=4.0.0
